test_lib: Un-inline expect_equal
This is done because the compiler is especially bad at lazily setting up
the function arguments for expect_equal fail. What ends up happening is
that the code does all of the work and preparation to have all the
values in argument registers anyways, but we still inline the
comparison, which just resulted in strictly worse codegen.